+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 -----
 
 * Add the `firstname` and `lastname` fields to a user
+* The `name` of a tenant is now optional
 * A new resource has been added to list a user external authencation methods
 
   * GET `0.1/users/<user_uuid>/external`